+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.12.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.19.0
+
 ## 0.12.1
 
 ### Patch Changes