+++ packages/@runlightyear/openai/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/openai
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.19.0
+
 ## 0.3.2
 
 ### Patch Changes